+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, Keira & Sky'
 author = 'Keira & Sky'
 
 # The full version, including alpha/beta/rc tags
-release = '0.3.1'
+release = '0.4.0'
 
 
 # -- General configuration ---------------------------------------------------